Extract discussion flow handlers
