+++ dev-docs/module-reference.md
@@ -478,7 +478,7 @@ Data export, import, and reset.
 
 ### `chat.js`
 
-Chat public barrel and startup entry point. Importing this module installs `chat-window-bindings.js` for legacy inline handlers, then re-exports the public chat helpers from the feature modules. Per-marker/correlation prompt builders live in `chat-marker-prompts.js`; direct send/streaming and image send integration live in `chat-send.js`; chat transcript rendering lives in `chat-render.js`; empty/onboarding message states live in `chat-empty-state.js`; chat-first onboarding handlers and provider quiz helpers live in `chat-onboarding.js`; multi-persona discussion flows live in `chat-discussion.js`, with callback bridging in `chat-discussion-callbacks.js`, persona/thread state helpers in `chat-discussion-state.js`, round execution in `chat-discussion-round-runner.js`, round prompt helpers in `chat-discussion-round-prompts.js`, round API request helpers in `chat-discussion-round-request.js`, thread-bound round persistence in `chat-discussion-round-state.js`, live round message DOM helpers in `chat-discussion-round-view.js`, and Discuss button/picker/continue controls in `chat-discussion-ui.js`; panel chrome lives in `chat-panel.js`; FAB nudge state lives in `chat-nudge.js`; personality selection and custom persona editing live in `chat-personalities.js`; current-thread history persistence lives in `chat-history.js`; message action bars live in `chat-actions.js`; chat image attachment state lives in `chat-images.js`; thread index storage and rail rendering live in `chat-threads.js`; thread rail message search and match highlighting live in `chat-thread-search.js`.
+Chat public barrel and startup entry point. Importing this module installs `chat-window-bindings.js` for legacy inline handlers, then re-exports the public chat helpers from the feature modules. Per-marker/correlation prompt builders live in `chat-marker-prompts.js`; direct send/streaming and image send integration live in `chat-send.js`; chat transcript rendering lives in `chat-render.js`; empty/onboarding message states live in `chat-empty-state.js`; chat-first onboarding handlers and provider quiz helpers live in `chat-onboarding.js`; multi-persona discussion exports live in `chat-discussion.js`, with flow handlers in `chat-discussion-flow.js`, callback bridging in `chat-discussion-callbacks.js`, persona/thread state helpers in `chat-discussion-state.js`, round execution in `chat-discussion-round-runner.js`, round prompt helpers in `chat-discussion-round-prompts.js`, round API request helpers in `chat-discussion-round-request.js`, thread-bound round persistence in `chat-discussion-round-state.js`, live round message DOM helpers in `chat-discussion-round-view.js`, and Discuss button/picker/continue controls in `chat-discussion-ui.js`; panel chrome lives in `chat-panel.js`; FAB nudge state lives in `chat-nudge.js`; personality selection and custom persona editing live in `chat-personalities.js`; current-thread history persistence lives in `chat-history.js`; message action bars live in `chat-actions.js`; chat image attachment state lives in `chat-images.js`; thread index storage and rail rendering live in `chat-threads.js`; thread rail message search and match highlighting live in `chat-thread-search.js`.
 
 **Key exports:**
 - `sendChatMessage()` — re-exported from `chat-send.js` for existing callers
@@ -573,11 +573,19 @@ Chat-first onboarding helpers. Owns the provider quiz render helper, onboarding 
 
 ### `chat-discussion.js`
 
-Multi-persona discussion/debate orchestration. Owns manual-message discussion turns, discussion start/continue/end flow, and picker decisions. It delegates callback bridging to `chat-discussion-callbacks.js`, Discuss button/picker/continuation prompt DOM to `chat-discussion-ui.js`, per-persona round execution to `chat-discussion-round-runner.js`, and thread-bound round persistence to `chat-discussion-round-state.js`; it re-exports persona/thread state helpers from `chat-discussion-state.js`.
+Multi-persona discussion public barrel. Re-exports discussion flow handlers from `chat-discussion-flow.js`, callback configuration from `chat-discussion-callbacks.js`, UI compatibility helpers from `chat-discussion-ui.js`, and persona/thread state helpers from `chat-discussion-state.js`.
 
 **Key exports:** `configureChatDiscussion`, `getThreadPersonaCount`, `updateDiscussButton`, `getCurrentDiscussionState`, `sendDiscussionUserTurn`, `restoreDiscussionContinuePrompt`, `showDiscussContinuePrompt`, `removeDiscussContinuePrompt`, `cleanupDiscussionState`, `startDiscussion`, `startDiscussionFromPicker`, `continueDiscussion`, `endDiscussion`
 
 **Window exports:** assigned by `chat-window-bindings.js` for existing inline handlers and `chat-threads.js` callbacks.
+
+### `chat-discussion-flow.js`
+
+Multi-persona discussion flow handlers. Owns manual-message discussion turns, active-discussion restore, continuation prompt persistence, cleanup/end lifecycle, picker decisions, single-persona add-on turns, and discussion round completion handoff. It delegates picker/prompt DOM to `chat-discussion-ui.js`, per-persona execution to `chat-discussion-round-runner.js`, and thread-bound persistence to `chat-discussion-round-state.js`.
+
+**Key exports:** `sendDiscussionUserTurn`, `restoreDiscussionContinuePrompt`, `showDiscussContinuePrompt`, `cleanupDiscussionState`, `startDiscussion`, `startDiscussionFromPicker`, `continueDiscussion`, `endDiscussion`
+
+**Window exports:** none directly; `chat-discussion.js` re-exports the public handlers and `chat-window-bindings.js` assigns the compatibility handlers.
 
 ### `chat-discussion-callbacks.js`
