+++ tests/test-correctness-phase2.js
@@ -136,6 +136,7 @@ const pwaAppShellAssets = [
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
   '/js/chat-discussion-callbacks.js',
+  '/js/chat-discussion-flow.js',
   '/js/chat-discussion-round-runner.js',
   '/js/chat-discussion-round-prompts.js',
   '/js/chat-discussion-round-request.js',
